requirements.txt
Django==3.1
djangorestframework==3.11.1
celery==4.4.7
redis==3.5.3
PyYAML==5.3.1
ua-parser==0.10.0
user-agents==2.2.0
django-user-agents==0.4.0
